requirements.txt
torch
opencv-python
pillow
fastapi
pydantic
ultralytics
numpy
matplotlib
tqdm
transformers
scikit-learn
uvicorn